requirements.txt
requests
python-dotenv
schedule
stripe
discord.py
langchain
langchain-openai
openai